Attempt to resolve incompatible versions
Trying to install all of the following:
* boto3
* aiobotocore
* botocore
* pandas-redshift
* snowflake-sqlalchemy
results in pipenv reporting incompatible versions in the resolved dependencies. Even poetry, with it's supposedly better dependency resolver, can't resolve the problem:
```
Using version ^1.12.39 for boto3
Using version ^1.0.2 for aiobotocore
Using version ^1.15.39 for botocore
Using version ^2.0.3 for pandas-redshift
Using version ^1.2.3 for snowflake-sqlalchemy
...
[InvalidRequirement]
```
Looking at the dependency graph with doing a `pipenv install --skip-lock` first, the following dependencies are found (leaving out the extras):
```
aiobotocore==1.0.2
  - botocore [required: >=1.15.32,<1.15.33, installed: 1.15.39]
pandas-redshift==2.0.3
  - boto3 [required: Any, installed: 1.11.17]
    - botocore [required: >=1.14.17,<1.15.0, installed: 1.15.39]
    - s3transfer [required: >=0.3.0,<0.4.0, installed: 0.3.3]
      - botocore [required: >=1.12.36,<2.0a.0, installed: 1.15.39]
  - pandas [required: Any, installed: 1.0.3]
snowflake-sqlalchemy==1.2.3
  - snowflake-connector-python [required: <3.0.0, installed: 2.2.3]
    - boto3 [required: >=1.4.4,<1.12, installed: 1.11.17]
      - botocore [required: >=1.14.17,<1.15.0, installed: 1.15.39]
      - s3transfer [required: >=0.3.0,<0.4.0, installed: 0.3.3]
        - botocore [required: >=1.12.36,<2.0a.0, installed: 1.15.39]
    - botocore [required: >=1.5.0,<1.15, installed: 1.15.39]
```
Numbering the botocore dependencies from top to bottom as 1 through 6, we find that 2, 3, 4, 5, and 6 are mutually compatible, but 1 and 2 are not, nor are 1 and 6. This is an attempt to resolve the issue with as little work as possible (but may fail due to how tightly coupled aiobotocore is with botocore).