Attempt to resolve incompatible versions
Trying to install aiobotocore and pandas_redshift does not work. Neither pipenv nor poetry can find a solution, because aiobotocore constrains the botocore in a manner that doesn't let it work.

```
aiobotocore==1.0.4
  - botocore [required: >=1.15.32,<1.15.33, installed: 1.16.1]
pandas-redshift==2.0.3
  - boto3 [required: Any, installed: 1.13.1]
    - botocore [required: >=1.16.1,<1.17.0, installed: 1.16.1]
    - s3transfer [required: >=0.3.0,<0.4.0, installed: 0.3.3]
      - botocore [required: >=1.12.36,<2.0a.0, installed: 1.16.1]
```
botocore can't be both < 1.15.33 and > 1.16.1.

This is a hack of the required package version numbers in setup.py in hopes of letting pipenv find a solution.